_requirements_db_env_commented.txt
# ========================
# Core NLP & Embedding Stack
# ========================
sentence-transformers==2.2.2       # Sentence embedding models (e.g., all-MiniLM-L6-v2)
transformers==4.25.1               # Hugging Face transformer models
huggingface-hub==0.10.1            # Hugging Face model hosting & downloading
tokenizers==0.13.3                 # Fast tokenization (used by transformers)
sentencepiece==0.2.0               # Required by some multilingual models

# ========================
# Machine Learning & Math
# ========================
scikit-learn==1.6.1                # Clustering, classification, and other ML utils
scipy==1.15.2                      # Scientific computing and linear algebra
numpy==2.2.4                       # Core numerical operations
joblib==1.4.2                      # Parallel processing (used by scikit-learn)
threadpoolctl==3.6.0               # Thread control (used by scikit-learn)
sympy==1.13.1                      # Symbolic mathematics (possibly unused, optional)
mpmath==1.3.0                      # Arbitrary-precision arithmetic (dependency of sympy)

# ========================
# Vector Search
# ========================
faiss-cpu==1.10.0                  # Fast vector similarity search (CPU-only build)

# ========================
# PyTorch and NVIDIA Backend
# ========================
torch==2.6.0                       # Core deep learning library
torchvision==0.21.0                # Image models & transformations
triton==3.2.0                      # PyTorch backend optimization
# The following are CUDA 12 runtime dependencies (used for GPU acceleration)
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127

# ========================
# Data Processing & Utilities
# ========================
nltk==3.9.1                        # Tokenization, stopwords, and other NLP tools
regex==2024.11.6                  # Better regex support (used by tokenizers/NLP)
fsspec==2025.3.2                  # Filesystem interface abstraction (used by transformers)
filelock==3.18.0                  # Used for managing parallel downloads
PyYAML==6.0.2                     # Config file handling
packaging==24.2                   # Version checking and packaging utils
typing_extensions==4.13.1         # Forward compatibility for typing features
Jinja2==3.1.6                     # Templating engine (possibly used by something like transformers)
MarkupSafe==3.0.2                 # Dependency of Jinja2

# ========================
# Networking / HTTP
# ========================
requests==2.32.3                  # HTTP requests (used by transformers & HF hub)
urllib3==2.3.0                    # Lower-level HTTP handling
idna==3.10                        # Internationalized domain names (used by requests)
certifi==2025.1.31                # Root certificate authority list
charset-normalizer==3.4.1         # Text encoding detection for HTTP responses

# ========================
# Other Tools
# ========================
tqdm==4.67.1                      # Progress bars
click==8.1.8                      # CLI tool support
psycopg2-binary==2.9.10           # PostgreSQL database adapter
pillow==11.1.0                    # Image loading (used by torchvision)
safetensors==0.5.3                # Safe, fast tensor serialization (used by HF models)
